+++ config/elastic.exs
@@ -1,10 +1,5 @@
 import Config
 
-# config :td_qx, TdDd.DataStructures.Search,
-#   es_scroll_size: 10_000,
-#   es_scroll_ttl: "1m",
-#   max_bulk_results: 100_000
-
 config :td_core, TdCore.Search.Cluster,
   # The default URL where Elasticsearch is hosted on your system.
   # Will be overridden by the `ES_URL` environment variable if set.
@@ -17,40 +12,15 @@ config :td_core, TdCore.Search.Cluster,
 
   # The library used for JSON encoding/decoding.
   json_library: Jason,
-  default_options: [
-    timeout: 5_000,
-    recv_timeout: 40_000
-  ],
   aliases: %{
     quality_controls: "quality_controls"
-  },
+  }
 
-  # Aggregations default
-  aggregations: %{
-    "domain" => 50,
-    "user" => 50,
-    "system" => 50,
-    "default" => 50
-  },
-  default_settings: %{
-    "number_of_shards" => 5,
-    "number_of_replicas" => 1,
-    "refresh_interval" => "5s",
-    "max_result_window" => 10_000,
-    "index.indexing.slowlog.threshold.index.warn" => "10s",
-    "index.indexing.slowlog.threshold.index.info" => "5s",
-    "index.indexing.slowlog.threshold.index.debug" => "2s",
-    "index.indexing.slowlog.threshold.index.trace" => "500ms",
-    "index.indexing.slowlog.level" => "info",
-    "index.indexing.slowlog.source" => "1000",
-    "index.mapping.total_fields.limit" => "3000"
-  },
-  indexes: %{
-    quality_controls: %{
+config :td_core, TdCore.Search.Cluster,
+  indexes: [
+    quality_controls: [
       store: TdQx.Search.Store,
       sources: [TdQx.QualityControls.QualityControl],
-      bulk_page_size:
-        System.get_env("BULK_PAGE_SIZE_QUALITY_CONTROLS", "5000") |> String.to_integer(),
       bulk_wait_interval: 0,
       bulk_action: "index",
       settings: %{
@@ -74,5 +44,5 @@ config :td_core, TdCore.Search.Cluster,
           }
         }
       }
-    }
-  }
+    ]
+  ]