+++ config/test.exs
@@ -20,6 +20,7 @@ config :td_qx, TdQxWeb.Endpoint, server: false
 config :td_cluster, TdCluster.ClusterHandler, MockClusterHandler
 
 config :td_core, TdCore.Search.Cluster, api: ElasticsearchMock
+config :td_core, TdCore.Search.IndexWorker, TdCore.Search.IndexWorkerMock
 
 # Print only warnings and errors during test
 config :logger, level: :warning