+++ lib/td_qx_web/controllers/search_controller.ex
@@ -7,10 +7,10 @@ defmodule TdQxWeb.SearchController do
   alias TdQx.Executions.Actions
   alias TdQx.Permissions
   alias TdQx.QualityControls
+  alias TdQx.Search.Indexer
 
   action_fallback(TdQxWeb.FallbackController)
 
-  # @index_worker Application.compile_env(:td_qx, :dq_index_worker)
   @default_page 0
   @default_size 20
 
@@ -45,7 +45,7 @@ defmodule TdQxWeb.SearchController do
     claims = conn.assigns[:current_resource]
 
     with :ok <- Bodyguard.permit(QualityControl, :reindex, claims) do
-      Search.IndexWorker.reindex(:quality_controls, :all)
+      Indexer.reindex(:all)
       send_resp(conn, :accepted, "")
     end
   end