+++ test/support/mocks.ex
@@ -1,3 +1,2 @@
 Mox.defmock(ElasticsearchMock, for: Elasticsearch.API)
 Mox.defmock(MockClusterHandler, for: TdCluster.ClusterHandler)
-Mox.defmock(MockIndexWorker, for: TdCore.Search.IndexWorker)